Extract storage helpers
